Minor changes in the api.

+++ microraidenj/app/src/main/java/cm/aptoide/pt/ethereum/EthereumApi.java
@@ -8,7 +8,7 @@ import rx.Observable;
 
 public interface EthereumApi {
 
-  Observable<Long> getCurrentNonce(String address);
+  Observable<Integer> getCurrentNonce(String address);
 
   Observable<TransactionResultResponse> sendRawTransaction(String rawData);
